cifs: introduce cifs credential
Longhorn 3599
