+++ pkg/types/types.go
@@ -23,6 +23,9 @@ const (
 	AWSEndPoint  = "AWS_ENDPOINTS"
 	AWSCert      = "AWS_CERT"
 
+	CIFSUsername = "CIFS_USERNAME"
+	CIFSPassword = "CIFS_PASSWORD"
+
 	HTTPSProxy = "HTTPS_PROXY"
 	HTTPProxy  = "HTTP_PROXY"
 	NOProxy    = "NO_PROXY"
